Model Code/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
# Python version 3.9

Pillow == 8.4.0
anydex == 0.1.0
auto-python-2014 == 14.1.5
auto-python-for-mobile == 14.1.5
bayesbridge == 0.2.5
bitmex-ws == 0.5.0
bitmex-ws-onix == 0.3.1
blocked-matrix-utils == 1.0.9
bmcs == 0.0.2a46
build-py == 1.0.0
buildtools == 1.0.6
cgp-dss-data-loader == 1.1.0
clicmod == 0.7.10
cmdline-brute == 0.5
cninfowebapi == 0.1.4
consultor == 0.2.0
cryptools == 0.0.6
data == 0.4
ddns == 2.10.3
dials == 0.0.1
django-parakeet == 0.1
douyin-downloader == 1.0.4
feb == 0.1
flask-nsfw == 6
flatten-json == 0.1.13
gardener-cicd-base == 1.1620.0
gcsa == 1.3.0
jw == 1.0
numpy == 1.20.3
package-generator == 0.0.1
patchman == 2.0.3
playground-resumeparser == 0.0.1
plda == 1.0.0
prestoadmin == 2.12
pytorch_lightning == 1.5.9
rapier == 0.0.4
scipy == 1.7.1
sr-python-commons == 0.121
stackmob-cli == 0.1.1
stackmob-parse-migrator == 1.0.1
streamlink-helper == 0.6
swarm-war == 1.0.0b1
thctools == 1.0.3
torch == 1.10.1
torchdiffeq == 0.2.2
tuido == 0.0.3
utillib == 0.2
vcv == 3.0.0.1
yang0textutils == 0.0.1
